requirements_dev.txt
pip==19.0.1
bumpversion==0.5.3
wheel==0.32.3
watchdog==0.9.0
flake8==3.6.0
tox==3.7.0
coverage==4.5.2
Sphinx==1.8.3
cryptography==2.5
PyYAML==3.13
pytest==4.1.1
requests==2.21.0
Pillow==5.4.1
